Fix data types and add success message.

+++ app/src/main/java/com/raudonikiss/weatherforecast/network/response/WeatherForecastResponseBody.kt
@@ -7,12 +7,12 @@ data class WeatherForecastResponseBody(
     val clouds: Clouds,
     val cod: Int,
     val coord: Coord,
-    val dt: Int,
+    val dt: Long,
     val id: Int,
     val main: Main,
     val name: String,
     val sys: Sys,
-    val visibility: Int,
+    val visibility: Double,
     val weather: List<Weather>,
     val wind: Wind
 ){
